Map uniquely the original (non-filtered) reads on the assembly with LAST.
In comparison with the other aligners, the combination of lastal and
last-split searches for the optimal set of local alignments of one read
to a genome.  Thus it is a one-to-many alignment (a read can hit
multiple contigs).

This data can be used to compute or display coverage.  There are pros
and cons for using the non-filtered reads, but at the moment I plan to
also use them to purge haplotypes.  Both kind of alignments can be
generated later if needed.